Tarefa1

Description

Este é o módulo correspondente à Tarefa 1 do trabalho que, através de diversas funções, desenvolve um mecanismo para gerar labirintos, a base do jogo.

Documentation

type Maze = [Corridor]

Construtores que vão ser usados posteriormente ao longo desta e das restantes Tarefas

type Corridor = [Piece]

data Piece

Constructors

Food FoodType

Wall

Empty

data FoodType

Constructors

Big

Little

printMaze :: Maze -> String

Converte um Maze numa String

generateRandoms :: Int -> Int -> [Int]

Dada uma semente retorna a lista the números inteiros gerados aleatoriamente

randomNumbers :: Int -> Int

Dada uma semente retorna um inteiro gerado aleatoriamente

subList :: Int -> [Int] -> [[Int]]

Converte uma lista numa lista de lista de tamanho n

convertsPiece :: Int -> Piece

Converte um número inteiro numa Piece 3 = Food Big 0 n < 70 <= Food Little 70 <= 99 <= Wall

printCorridor :: Corridor -> String

Converte um Corridor numa string

convertsCorridor :: [Int] -> Corridor

Converte uma lista de inteiros num Corridor

convertsMaze :: [[Int]] -> Maze

Converte uma lista de listas de inteiros num Maze

innerZone :: Int -> Int -> Int -> Maze

Dados três números inteiros devolve um Maze com a parte interior do Maze inicial

allWalls :: Int -> [Piece]

Converte um número inteiro numa lista de Piece Wall

middleHeight :: Maze -> Int

Converte um Maze num inteiro sendo este último a altura do Maze

middleCorridors :: Maze -> [Corridor]

Converte um Maze numa lista de Corridor sendo estes a parte central do Maze

breakWall :: [Corridor] -> [Corridor]

Dada uma lista de Corridors devolve outra mesma lista sem as Wall do início e fim de cada Corridor

openTunnel :: Maze -> Maze

Converte um Maze num outro Maze com um túnel central

mazeWithHome :: [Corridor] -> Maze

Converte uma lista de Corridor num Maze com a casa dos fantasmas já implementada no centro do mesmo

buildHome :: [Corridor] -> [Corridor]

Dada uma lista de Corridor devolve outra lista de Corridor que representa apenas a casa dos fantasmas

corridorsLarge :: Maze -> Int

Converte um Maze num inteiro sendo este último a largura do Maze

evenBuildHome1 :: [Piece]

Funções auxiliares para a construção de cada Corridor da casa dos fantasmas sendo a largura par

evenBuildHome2 :: [Piece]

evenBuildHome3 :: [Piece]

evenBuildHome4 :: [Piece]

evenBuildHome5 :: [Piece]

oddBuildHome1 :: [Piece]

Funções auxiliares para a construção de cada Corridor da casa dos fantasmas sendo a largura ímpar

oddBuildHome2 :: [Piece]

oddBuildHome3 :: [Piece]

oddBuildHome4 :: [Piece]

oddBuildHome5 :: [Piece]

generateMaze :: Int -> Int -> Int -> Maze

Dados dois inteiros e uma semente retorna um Maze com a devida estrutura

displayMaze :: Maze -> IO ()

Apresenta o Maze com a sua devida estrutura