Tarefa2

Description

Este é o módulo correspondente à Tarefa 2 do trabalho que, através de diversas funções, realiza jogadas para os jogadores do tipo Pacman.

Documentation

whatPiece :: Coords -> Maze -> Piece

Dadas as Coords e um Maze retorna a Piece que se encontra naquela posição

play :: Play -> State -> State

Função objetivo que dada uma jogada (Play) e um Starte altera o State de um jogador

ordPlayers :: [Player] -> [Player]

Dada uma lista de Player retorna a mesma lista com os Player devidamente ordenados

testIfPacGho :: Player -> Bool

Dado um Player verifica se este é do tipo Ghost

newParameters :: Maze -> Play -> [Player] -> [Player]

Dado um Maze, uma jogada (Play) e um jogador retorna uma lista de Player com as devidas alterações

loseLives :: Player -> Player

Dado um Player do tipo Pacman retorna o mesmo com menos uma vida e altera o seu modo

takeGhost :: Maze -> Coords -> [Player] -> [Player]

Dado um Maze, Coords e uma lista de Player do tipo Ghost retorna esta última com as alterações no seu modo enviando-o para a parte central do Maze

centerTunnel :: Maze -> Coords

Converte um Maze em Coords relativas ao centro do túnel

killGhost :: [Player] -> [Player]

Dada uma lista de Player do tipo Ghost retorna o mesmo com o seu modo Dead

badGhost :: Coords -> [Player] -> Piece

Dadas Coords e uma lista de Players do tipo Ghost retorna um Piece que vai prejudicar o Pacman

deliciousGhost :: Coords -> [Player] -> Piece

Dadas Coords e uma lista de Players do tipo Ghost retorna um Piece que vai favorecer o Pacman

y'sMaze :: Maze -> Int

Converte um Maze num inteiro sendo este último a largura do Maze

lMaze :: Maze -> Int

Converte um Maze num inteiro sendo este último a altura do Maze