Tarefa3

Description

Este é o módulo correspondente à Tarefa 3 do trabalho que, dado um labirinto válido, converte-o numa sequência de instruções de tornando-o mais compacto.

Documentation

compactMaze :: Maze -> Instructions

Dado um Maze retorna Instructions que representam o Maze de uma forma mais compacta

auxMain :: [Corridor] -> [Instruction]

Converte uma lista de Corridor numa lista de Instruction

subaux :: [Instruction] -> [Instruction]

Converte uma lista de Instruction numa mesma em que se há Corridor iguais aparecerá Repeat x (sendo x a posição do Corridor inicial)

joinInstruction :: Corridor -> Instruction

Dado um Corridor retorna a sua Instruction correspondente

givesCorridor :: Corridor -> [(Int, Piece)]

Converte um Corridor numa lista de pares de (Int,Piece)

acumulatePiece :: [(Int, Piece)] -> [(Int, Piece)]

Dada uma lista de (Int,Piece) retorna uma semelhante em que, se as Piece consecutivas forem iguais vai-se acumulando o número das mesmas

samePiece :: Piece -> Piece -> Bool

Dadas duas Piece verifica se estas são iguais

instructPosition :: Instruction -> Instructions -> Int

Compara se uma Instruction é igual à cabeça das Instructions e assim, devolve a posição dessa Instruction inicial

sameInstruction :: Instruction -> Instruction -> Bool

Dadas duas Instruction verifica se estas são iguais

sizeInstructions :: Instructions -> Int

Testa o nível de compactação da função