Tarefa4

Description

Este é o módulo correspondente à Tarefa 4 do trabalho que, através de diversas funções, permite determinar as consequências da passagem de um instante de tempo num estado do jogo.

Documentation

defaultDelayTime :: Integer

Introdução

Esta tarefa foi a qual se desenvolveu para que fosse possível "movimentar" todo o tipo de jogadores no labirinto do jogo. Passado um instante de tempo, há uma atualização no estado do jogo mas, também, nos requerimentos necessárias para ser possível efetuar uma jogada. Nesta parte do projeto é inserido o contexto de uma nova biblioteca: NCurses.

Objetivos

O principal objetivo que deve ser alcançado nesta tarefa é o de que, dado um step e o estado atual do jogo, este sofra uma atualização. A atualização a aplicar deve ter em conta as jogadas de cada tipo de jogadores presentes. A cada instante de tempo que passa é possível determinar a resposta a cada situação com que o jogador se depare. A função Play, implementada da Tarefa 2, passa assim a ter mais exigências

Abertura e fecho da boca do Pacman a cada jogada;

Perda de tempo Mega a cada jogada (quando o modo do Pacman é Mega);

Quando se dá o término do tempo Mega, o Pacman volta ao modo Alive;

Se todos os Pacman estão em modo Normal então, os Fantasmas devem voltar ao modo Alive;

Dependendo da velocidade do jogador e da paridade da repetição, estes devem progredir x jogadas.

Discussão e Conclusão

A integração da nova biblioteca NCurses trouxe alguma complexidade ao trabalho e, de tal forma, foi possível associar todas as ocorrências de jogada em jogada. A estratégia que adotamos para o desenvolvimento desta tarefa passou por se recorrer à função Play (função relativa às jogadas de qualquer jogador) e, também, às funções updateTime, nextFrame e updateControlledPlayer (inseridas no Main da 2ª fase) que controlam o tempo de jogo, fazem a atualização do estado atual do mesmo e permitem a intervenção do jogador através de comandos (teclas com setas). Infelizmente, não foi possível atingir o objetivo final pois o jogador do tipo Ghost não sofre qualquer alteração, ou seja, não se movimenta, ao contrário do Pacman.

Concluindo, não conseguimos atingir o resultado pretendido mas, através do desenvolvimento desta tarefa, tivemos a oportunidade de visualizar jogo praticamente completo e de uma maneira mais interativa.

passTime :: Int -> State -> State

Função principal da Tarefa 4

Dado um Inteiro (Step) e um State devolve um novo State atualizado, onde todos os joadores efeturaram uma jogada

l' :: [Int]

Lista que representa a numeração dos Steps pares

l'' :: [Int]

Lista que representa a numeração dos Steps ímpares

funApp' :: Maze -> [Player] -> [Player]

Dado um Maze e uma lista de Players devolve uma lista de Players com a função play' aplicada a cada um deles, esta função é ativada para Steps pares

play' :: Maze -> Player -> Player

Dado um Maze e um Player é efetuada nesse Player a função playPac ou a PlayGho, dependendo do Player, um certo número de vezes, conforme a velocity do Player

funApp'' :: Maze -> [Player] -> [Player]

Dado um Maze e uma lista de Players devolve uma lista de Players com a função play'' aplicada a cada um deles, esta função é ativada para Steps ímpares

play'' :: Maze -> Player -> Player

Dado um Maze e um Player é efetuada nesse Player a função playPac ou a PlayGho, dependendo do Player, um certo número de vezes, conforme a velocity do Player

playPac :: Maze -> Player -> Player

Dado um Maze e um Player (Pacman) efetua uma jogada nesse Player

playPacOtherwise :: Maze -> Player -> Player

Dado um Maze e um Player (Pacman) efetua uma jogada nesse Player

playGho :: Maze -> Player -> Player

Dado um Maze e um Player (Ghost) efetua uma jogada nesse Player

rightPiece :: Maze -> Player -> Piece

Dado um Maze e um Player devolve a peça à direita do Player

leftPiece :: Maze -> Player -> Piece

Dado um Maze e um Player devolve a peça à esquerda do Player

underPiece :: Maze -> Player -> Piece

Dado um Maze e um Player devolve a peça abaixo do Player

downPiece :: Maze -> Player -> Piece

Dado um Maze e um Player devolve a peça acima do Player

isPac :: Player -> Bool

Dado um Player diz-nos se é Pacman quando é True

getVel :: Player -> Double

Dado um Player devolve-nos a sua Velocity

whatPiece :: Coords -> Maze -> Piece

Dadas umas Coords e um Maze devolve-nos a Piece com essas Coords no Maze

nCoords :: Coords -> Player -> Player

Dadas umas Coords e um Player devolve o Player com novas Coords

addP :: Int -> Player -> Player

Dado um inteiro e um Player devolver o Player com novos Points resultantes da soma dos anteriores com o inteiro

main1 :: [Player] -> [Player]

Dada uma lista de Players devolve uma lista de Players que resulta da aplicação de funções à lista inicial

loseTmega :: [Player] -> [Player]

Dada uma lista de Players devolve uma nova lista que resulta de aplicar uma função aos Pacmans da lista inicial atualizando-lhes o Tempo e o Modo em Mega

takeTmega :: Player -> Player

Dado um Player devolve o mesmo Player com menos Tempo Mega, se o Tempo Mega é Zero o Player passa ao Modo Normal

separatePlayers :: [Player] -> ([Player], [Player])

Dada uma lista de Players devolver um tuplo cuja primeira lista contém Pacmans e a segunda Ghosts

testMega :: [Player] -> Bool

Dada uma lista de Players devolve True se a lista contém Pacmans em Modo Mega, caso contrário, False

sayIfMega :: Player -> Bool

Dado um Player devolve True se o Modo for Mega, caso contrário, False

killGhost :: [Player] -> [Player]

Dada uma lista de Players Ghost devolve uma lista com todos esses Ghosts em Modo Dead

pMega :: Player -> Player

Dado um Player qualquer que seja o Modo devolve o Player no Modo Mega

ifGhoDead :: ([Player], [Player]) -> [Player]

Dada um tuplo, cujas componentes sao listas de Players devolve uma lista de Players

testGhoDeadList :: [Player] -> Bool

Dada uma lista de Ghosts devolve True, se há algum em Modo Dead, caso contrário, devolve False

eatGho :: ([Player], [Player]) -> [Player]

Dado um tuplo com listas de Players devolve uma lista de Players

findGhoDead :: Player -> [Player] -> Player

Dado um Player e uma lista de Ghosts acrescenta dez Points ao Player se as Coords coincidem com as de um Ghost Dead

compareC :: Player -> Player -> Bool

ifGhoAlive :: ([Player], [Player]) -> [Player]

Dada um tuplo, cujas componentes sao listas de Players devolve uma lista de Players

testGhoAliveList :: [Player] -> Bool

Dada uma lista de Ghosts devolve True, se há algum em Modo Alive, caso contrário, devolve False

losePoints :: ([Player], [Player]) -> [Player]

Dado um tuplo com duas lista de Players devolve uma lista de Players que resulta de concatenar ambas após aplicar funções à primeira

findGhoAlive :: Player -> [Player] -> Player

Dado um Player e uma lista de Ghosts remove Lives ao Player se as Coords coincidem com as de um Ghost Alive

takeGhost :: Maze -> Player -> Player

Dado um Maze e um Player (Ghost) leva esse Player para o centro do Maze

loseLives :: Player -> Player

Dado um Pacman retira-lhe uma Live sempre que este tem até zero vidas, quando este já tem zero muda-lhe o Modo para Dying

centerTunnel :: Maze -> Coords

Dado um maze devolve as Coords do centro do Maze

pacMode :: Player -> PacMode

Dado um Player do tipo Pacman devolve o seu Modo

ghoMode :: Player -> GhostMode

Dado um Player do tipo Ghost devolve o seu Modo

getC :: Player -> Coords

Dado um Player devolve as suas Coords

getMouth :: Player -> Mouth

Dado um Player devolve o Estado da Mouth

closedMouth :: Player -> Player

Dado um Player muda-lhe o Estado da Mouth para Open se estiver Closed e vice-versa

moveTo :: (Int, Int) -> Player -> Player

Dado um par de inteiros e um Player devolve-nos as novas Coords do Player que resulta de adicionar às iniciais as componentes do par

sndC :: Player -> Int

Dado um Player devolve a segunda componente das Coords desse Player

y'sMaze :: Maze -> Int

Dado um Maze de devolve o comprimento dos corredores do Maze

lMaze :: Maze -> Int

Dado um Maze devolve a altura do Maze

changeSnd :: Int -> Player -> Player

Dado um inteiro e um Player é devolvido o mesmo Player agora com a segunda componente da sua Coord correspondente ao inteiro dado