Tarefa5

Description

Este é o módulo correspondente à Tarefa 5 do trabalho que, através de diversas funções, permite a movimentação dos fantasmas no jogo.

Documentation

ghostPlay :: State -> [Play]

Introdução

Nesta tarefa tinha que se desenvolver as jogadas e o comportamento dos fantasmas no jogo. A tarefa 2 é aqui usada diretamente como auxílio pois é a que retrata as jogadas possíveis neste projeto.

Objetivos

São diversos os objetivos a alcançar tais que, a função principal a implementar recebe o estado do jogo e devolve as várias jogadas dos fantasmas ativos no jogo. Estas jogadas devem ser aplicadas de maneira a que seja aplicada a melhor estratégia para atuar perante o seu "adversário" Pacman. Tem que se ter em consideração ao modo (Alive ou Dead) do fantasma pois perante isto, o mesmo deve ter respostas diferentes, se não mesmo opostas. Assim, as estratégias do fantasma baseiam-se em:

(modo Alive) ir atrás do Pacman (através das suas coordenadas) e posicionar-se quatro espaços à frente do mesmo (dependendo da sua orientação);

(modo Dead) fugir do Pacman e sempre que se encontra com uma parede move-se para a sua direita, no sentido dos ponteiros de relógio.

Discussão e Conclusão

O procedimento usado no desenvolvimento desta tarefa passou pela integração de algumas funções que foram necessárias na tarefa 2 (da 1ª fase) e, com estas, incorporar o potencial comportamento dos Ghost. Para isto, era necessário o estado atual do jogo que, com as devidas alterações, apresentar-se-ia uma lista de jogadas sendo estas a melhor alternativa. Tendo em conta o modo do Pacman (Normal ou Mega) e o modo do Ghost (Alive ou Dead), a função principal passava pelo planeamento da fuga do Ghost ou, então, pela corrida atrás do Pacman. Através do método de Manhattan consegui-se calcular a distância entre os jogadores adversários e assim, começaria o processo de perseguição.

Por fim, consideramos que demos o nosso melhor para atingir o objetivo final e, esta tarefa permitiu perceber ainda melhor todo o conceito de circulação dos jogadores, tanto Pacman como Ghost, e o possível manipulamento no jogo.

Função principal da Tarefa 5

Dado um State devolve uma lista de jogadas a efetuar pelos Ghost

chaseScatter :: State -> ([Player], [Player]) -> [Play]

Funçao que dado um State e um tuplo com listas de Pacman e Ghost, respetivamente, devolve uma lista de Play

chaseMode :: State -> Int -> Play

Função que dado um State um ID devolve uma play

scatterMode :: State -> Int -> Play

Função que dado um State um ID devolve uma play

metC :: Player -> Player -> Int

Dados dois Players devolve o resultado de aplicar a função methoDManhattan

methodManhattan :: Coords -> Coords -> Int

Dadas duas Coords devolve um inteiro (resultado da métrica de Manhattan entre as duas Coords)

minDist' :: Int -> [Player] -> Int

Dado um ID e a lista de Players esta função devolve a distância do jogador que se encontra mais próximo do Bot

funAux :: Int -> ([Player], [Player]) -> Int

Função que dado um inteiro e um tuplo com listas de Pacman e Ghost, respetivamente, devolve um inteiro

findById :: Int -> [Player] -> Player

Dado um ID e uma lista de Players devolve o Player com esse ID

removeById :: Int -> [Player] -> [Player]

Dado um ID e uma lista de Players devolve a lista sem o Player com o ID correspondente ao dado

play :: Play -> Maze -> Player -> Player

Dado um Play, um Maze e um Player, se a orientação do Player coincide com a do Play efetua uma jogada, caso contrário, muda a orientação para a do Play

goDead :: Maze -> Player -> Player

Dado um Maze e um Player devolve o Player atualizado

goAlive :: Maze -> Player -> Player

Função que dado um Maze e um player devolve esse Player atualizado

playOtherwise' :: Maze -> Player -> Player

Dado um Maze e um Player devolve o Player já movido

main2 :: [Player] -> [Player]

Dada uma lista de Players devolve uma lista de Players que resulta da aplicação de funções à lista inicial

turnGhost :: Player -> Player

Dado um Player é devolvido o Player com a direção oposta à inicial

orientation' :: Player -> Orientation

Dado um Player dá a sua Orientação

changeOrientation :: Orientation -> Player -> Player

Dado uma Orientation e um Player dá ao Player a nova Orientation