Tarefa6

Description

Este é o módulo correspondente à Tarefa 6 do trabalho que, através de diversas funções, é possível recorrer a um robô que jogue de uma maneira automática.

Documentation

bot :: Int -> State -> Maybe Play

Introdução

Esta Tarefa 6 consistiu na criação de um robô que, com os devidos requisitos, seja possível executar uma jogada. Também se dá uso à Tarefa 2 para se obter as diferentes jogadas, dadas pelo bot.

Objetivos

Com a passagem do tempo, é conhecido o estado do jogo pelo bot e este é que averigua se dá uso à função Play ou não, ou seja, pode escolher efetuar uma jogada ou não. A estratégia de decisão está de acordo com o State de um jogo válido e possível e o identificador do jogador.

Discussão e Conclusão

A ideia de um robô a fazer tudo sozinho parecia, inicialmente, complicado. Para se atingir o objetivo final, tomou-se como abordagem a realização de uma jogada para quando é possível calcular a distância entre o Pacman e o Ghost. Este cálculo baseia-se no método simples de Manhattan e podem ser interpretados jogadores de qualquer tipo. Para a concretização da jogada avalia-se a distância entre os adversários (como anteriormente dito), o modo do Pacman e, consequentemente, o modo do Ghost.

Em suma, acreditamos que fizemos tudo ao nosso alcance para chegar ao melhor resultado possível em que, o robô tenha o melhor desempenho esperado.

Função principal da Tarefa 6

Dado o id de um jogador e o estado do jogo é devolvido uma possível jogada

minDist :: Int -> [Player] -> Int

Dado um ID e a lista de Players esta função devolve a distância do jogador que se encontra mais próximo do Bot

lminDist :: Player -> [Player] -> [Player]

Dado um ID e a lista de Players esta função devolve a distância do jogador que se encontra mais próximo do Bot

final :: Play -> Maze -> [Player] -> [Player]

Dado um play, um Maze e uma lista de jogadores devolve a lista de jogadores com o Bot já atualizado

findBot :: Int -> [Player] -> (Player, [Player])

Dado um ID e uma lista de Players devolve um tuplo com o Player que é o Bot e com a restante lista dos players

playB :: Play -> Maze -> Player -> Player

Dado um Play, um Maze e um Player, se a orientação do Player coincide com a do Play efetua uma jogada, caso contrário, muda a orientação para a do Play

goahead :: Maze -> Player -> Player

Dado um Maze e um Player devolve o Player já movido

playOtherwise :: Maze -> Player -> Player

Dado um Maze e um Player devolve o Player já movido